refactor: Consolidate peer option components and remove redundant files
