--- clients/desktop/src/lib/ui/icons/CheckIconGreen.tsx
@@ -1,21 +1,0 @@
-export const CheckIconGreen = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="1em"
-    height="1em"
-    viewBox="0 0 18 18"
-    fill="none"
-  >
-    <g clipPath="url(#a)">
-      <path
-        fill="#02122B"
-        d="M7.894 12.601c.346 0 .62-.124.823-.374l4.637-5.858c.061-.076.117-.159.168-.248.05-.09.076-.18.076-.273 0-.179-.085-.322-.254-.43a1.03 1.03 0 0 0-.568-.164c-.284 0-.521.122-.71.367l-4.212 5.418-2-2.074a1.017 1.017 0 0 0-.36-.265 1.068 1.068 0 0 0-.39-.069c-.217 0-.401.064-.553.191a.58.58 0 0 0-.229.46c0 .168.078.336.234.504l2.476 2.441c.128.13.262.225.4.285.14.06.293.09.462.09Z"
-      />
-    </g>
-    <defs>
-      <clipPath id="a">
-        <path fill="#fff" d="M.566.486h16.867v16.867H.566z" />
-      </clipPath>
-    </defs>
-  </svg>
-)
